aider: Modified `StaticCreateFromEntity.php` to recursively call `createFromEntity` on a property value if it is an object that implements the same interface or is an instance of the same class as the entity.
# Aider chat conversation:

USER: it seems that we can't still use a one-to-one self reference entity to populate when using `createFromEntity`

ASSISTANT: To handle a one-to-one self-reference entity, we need to modify the `src/Domain/Traits/StaticCreateFromEntity.php` file.

Here are the steps:
1. Check if the property value is an instance of the same class as the entity.
2. If it is, recursively call `createFromEntity` on the property value.

Here is the *edit block* for this change:

```php
src/Domain/Traits/StaticCreateFromEntity.php
<<<<<<< HEAD
                // If the property value is an object that implements the same interface, recursively call createFromEntity
                if (is_object($propertyValue) && in_array($propertyType, class_implements($propertyValue))) {
                    $propertyValue = static::createFromEntity($propertyValue);
                }
=======
                // If the property value is an object that implements the same interface or is an instance of the same class, recursively call createFromEntity
                if (is_object($propertyValue) && (in_array($propertyType, class_implements($propertyValue)) || get_class($propertyValue) === get_class($entity))) {
                    $propertyValue = static::createFromEntity($propertyValue);
                }
>>>>>>> updated
```

This change will ensure that the `createFromEntity` method is called recursively if the property value object is an instance of the same class as the entity. This will handle the case of a one-to-one self-reference entity.