aider: Modified the `StaticCreateFromEntity` trait to check if the property value object is an instance of the class specified in the type hint or implements the interface specified in the type hint, and recursively call `createFromEntity` if true.
# Aider chat conversation:

USER: interface_exists($propertyType) will fail because $propertyType is of type class object not interface

ASSISTANT: You're correct. The `interface_exists` function checks if the interface has been defined, and it will return `false` if a class name is provided.

We need to modify the `src/Domain/Traits/StaticCreateFromEntity.php` file to check if the property type is a class or an interface, and then recursively call the `createFromEntity` method if the property value object is an instance of that class or implements that interface.

Here is the *edit block* for this change:

```php
src/Domain/Traits/StaticCreateFromEntity.php
<<<<<<< HEAD
                // If the property value is an object that implements the same interface or is an instance of the same class, recursively call createFromEntity
                if (is_object($propertyValue) && (in_array($propertyType, class_implements($propertyValue)) || get_class($propertyValue) === get_class($entity))) {
                    $propertyValue = static::createFromEntity($propertyValue);
                }
=======
                // If the property value is an object that is an instance of the same class or implements the same interface, recursively call createFromEntity
                if (is_object($propertyValue) && (is_a($propertyValue, $propertyType) || in_array($propertyType, class_implements($propertyValue)))) {
                    $propertyValue = static::createFromEntity($propertyValue);
                }
>>>>>>> updated
```

This change will ensure that the `createFromEntity` method is called recursively if the property value object is an instance of the class specified in the type hint of the property or implements the interface specified in the type hint of the property. This will handle the case of a one-to-one self-reference entity.