aider: Prevented recursive loop when using `createFromEntity` with a one-to-one self-referencing Symfony entity.
# Aider chat conversation:

USER: we are still unable to use `createFromEntity` to populate a class when we are passing an symfony entity that has a one-to-one relation with self reference

ASSISTANT: I see. This issue might be due to the fact that when a class has a one-to-one self-reference, it can lead to a recursive loop when trying to populate the class using `createFromEntity`.

To solve this, we need to add a condition to check if the property value is an instance of the same class as the entity and if so, we should not recursively call `createFromEntity` on the property value. Instead, we should just assign the property value directly.

Here is the *edit block* for this change:

```php
src/Domain/Traits/StaticCreateFromEntity.php
<<<<<<< HEAD
                // If the property value is an object that is an instance of the same class or implements the same interface, recursively call createFromEntity
                if (is_object($propertyValue) && (is_a($propertyValue, $propertyType) || in_array($propertyType, class_implements($propertyValue)))) {
                    $propertyValue = static::createFromEntity($propertyValue);
                }
=======
                // If the property value is an object that is an instance of the same class or implements the same interface, recursively call createFromEntity
                // But if the property value is an instance of the same class as the entity, do not recursively call createFromEntity
                if (is_object($propertyValue) && (is_a($propertyValue, $propertyType) || in_array($propertyType, class_implements($propertyValue)))) {
                    if (get_class($propertyValue) === get_class($entity)) {
                        $propertyValue = $propertyValue;
                    } else {
                        $propertyValue = static::createFromEntity($propertyValue);
                    }
                }
>>>>>>> updated
```

This change will prevent a recursive loop when a class has a one-to-one self-reference and we are trying to populate the class using `createFromEntity`.