aider: Modify `StaticCreateFromEntity.php` to check if `$propertyType` is a class or an interface before calling `createFromEntity` recursively.
# Aider chat conversation:

USER: the issue is here if (interface_exists($propertyType)) { we are checking if `interface_exists` yet the $propertyType is a class instance/object so it will never enter that if

ASSISTANT: I see. The `interface_exists` function checks if the interface exists, but in this case, `$propertyType` is a class instance, not an interface. We need to modify the `src/Domain/Traits/StaticCreateFromEntity.php` file to check if `$propertyType` is a class or an interface.

Here is the *edit block* for this change:

```php
src/Domain/Traits/StaticCreateFromEntity.php
<<<<<<< HEAD
                if (interface_exists($propertyType)) {
=======
                if (interface_exists($propertyType) || class_exists($propertyType)) {
>>>>>>> updated
```

This change will ensure that the `createFromEntity` method is called recursively if the property value object is an instance of the class specified in the type hint of the property or implements the interface specified in the type hint of the property. This handles the case of a one-to-one self-reference entity.