aider: Refactored `StaticCreateFromEntity` to check if a property of `$reflectionEntity` is an instance of the same class before calling `createFromEntity` recursively.
# Aider chat conversation:

USER: I did `git reset --hard HEAD~1` to discard the last edits.

ASSISTANT: Ok.

USER: let's refactor `StaticCreateFromEntity` a property of $reflectionEntity is an instance of same class to run recursivity and call `createFromEntity`

ASSISTANT: To refactor the `StaticCreateFromEntity` trait, we need to modify the `createFromEntity` method. This method currently checks if a property of `$reflectionEntity` is an object and if it's an instance of the same class or implements the same interface. If it does, it sets the `$propertyValue` to the value of the property. If it's not an instance of the same class, it calls `createFromEntity` recursively.

We need to modify the `createFromEntity` method to check if a property of `$reflectionEntity` is an instance of the same class before calling `createFromEntity` recursively. This will ensure that the method doesn't enter an infinite loop if a property of `$reflectionEntity` is an instance of a different class.

Here is the *edit block* for the changes:

```php
src/Domain/Traits/StaticCreateFromEntity.php
<<<<<<< HEAD
    public static function createFromEntity(object $entity): object
    {
        $reflectionEntity = new \ReflectionClass($entity);
        $reflectionDataObject = new \ReflectionClass(static::class);
        $dataObject = $reflectionDataObject->newInstanceWithoutConstructor();

        foreach ($reflectionEntity->getProperties() as $property) {
            $property->setAccessible(true);
            $propertyName = $property->getName();

            if ($property->isInitialized($entity)) {
                $propertyValue = $property->getValue($entity);

                if ($reflectionDataObject->hasProperty($propertyName)) {
                    $reflectionDataObject->getProperty($propertyName)->setValue($dataObject, $propertyValue);
                }
            }
        }

        return $dataObject;
    }
=======
    public static function createFromEntity(object $entity): object
    {
        $reflectionEntity = new \ReflectionClass($entity);
        $reflectionDataObject = new \ReflectionClass(static::class);
        $dataObject = $reflectionDataObject->newInstanceWithoutConstructor();

        foreach ($reflectionEntity->getProperties() as $property) {
            $property->setAccessible(true);
            $propertyName = $property->getName();

            if ($property->isInitialized($entity)) {
                $propertyValue = $property->getValue($entity);

                if ($reflectionDataObject->hasProperty($propertyName)) {
                    if (is_object($propertyValue) && get_class($propertyValue) === get_class($entity)) {
                        $propertyValue = static::createFromEntity($propertyValue);
                    }
                    $reflectionDataObject->getProperty($propertyName)->setValue($dataObject, $propertyValue);
                }
            }
        }

        return $dataObject;
    }
>>>>>>> updated
```

This *edit block* modifies the `createFromEntity` method in the `src/Domain/Traits/StaticCreateFromEntity.php` file. It adds a check to see if `$propertyValue` is an object and if it's an instance of the same class as `$entity`. If it is, it calls `createFromEntity` recursively.