aider: Implement dynamic loading of assets based on the installed path.
# Aider chat conversation:

USER: let's implement a new feature in webpack.config.js so that we would load the assets dynamically based on installed path

ASSISTANT: To implement dynamic loading of assets based on the installed path, we need to modify the `webpack.config.js` file. We will use Node's built-in `fs` module to read the directories and add entries dynamically.

Here is the *edit block* for the change:

```javascript
themes/zenchron/default-theme-child/webpack.config.js
<<<<<<< HEAD
    .addEntry('app_js', './assets/javascript/app.js')
    .addEntry('app_css', './assets/css/app.css')
=======
    const fs = require('fs');
    const path = require('path');

    const assetsDir = './assets';
    const assetTypes = ['javascript', 'css'];

    assetTypes.forEach((type) => {
        const dir = path.join(assetsDir, type);
        fs.readdirSync(dir).forEach((file) => {
            Encore.addEntry(`app_${type}_${file.split('.')[0]}`, path.join(dir, file));
        });
    });
>>>>>>> updated
```

This change will make Webpack scan the `assets/javascript` and `assets/css` directories and add an entry for each file it finds. The entry name will be `app_<type>_<filename>` where `<type>` is either `javascript` or `css` and `<filename>` is the name of the file without the extension.