aider: Added ExportAssetsFromThemeCommand to export assets from the current active theme to the root directory.
# Aider chat conversation:

ASSISTANT: To implement the `ExportAssetsFromThemeCommand`, we need to create a new file in the `src/Command` directory. This command will have a dependency on the `AssetManager` service, so we need to inject it into the constructor. We will define the command name as `app:export-assets-from-theme`. In the `execute` method, we will call the `AssetManager`'s `copyAssetsToActiveTheme` method.

Here is the *edit block* for the new `ExportAssetsFromThemeCommand`:

```php
/Users/m3tagh0st/code/ai-cms/src/Command/ExportAssetsToThemeCommand.php
<<<<<<< HEAD
=======
<?php

namespace App\Command;

use App\Service\AssetManager;
use App\Service\ThemeManager;
use Symfony\Component\Console\Command\Command;
use Symfony\Component\Console\Input\InputInterface;
use Symfony\Component\Console\Output\OutputInterface;

class ExportAssetsFromThemeCommand extends Command
{
    protected static $defaultName = 'app:export-assets-from-theme';

    private AssetManager $assetManager;
    private ThemeManager $themeManager;

    public function __construct(AssetManager $assetManager, ThemeManager $themeManager)
    {
        $this->assetManager = $assetManager;
        $this->themeManager = $themeManager;

        parent::__construct();
    }

    protected function configure()
    {
        $this
            ->setDescription('Exports assets from the current active theme to the root directory.')
            ->setHelp('This command allows you to export assets from the current active theme to the root directory...');
    }

    protected function execute(InputInterface $input, OutputInterface $output): int
    {
        $activeThemeName = $this->themeManager->getActiveThemeName();
        if ($activeThemeName) {
            $this->assetManager->copyAssetsToActiveTheme($activeThemeName);
            $output->writeln('Assets exported successfully from the active theme.');

            return Command::SUCCESS;
        }

        $output->writeln('No active theme found.');

        return Command::FAILURE;
    }
}
>>>>>>> updated
```

Please note that the `copyAssetsToActiveTheme` method assumes that there is a method in the `ThemeManager` service that can return the name of the current active theme. If such a method does not exist, we need to create it.