aider: Refactored Volume entity to use DTO and repository instead of direct usage of entities.
# Aider chat conversation:

USER: continue with the implementation, also we should use only the dto and repositorie without direct usage of entities from src/Entity

ASSISTANT: To implement the changes you've requested, we'll need to modify the following files:

1. src/Entity/Volume.php
2. src/Entity/Disk.php
3. src/Entity/File.php
4. src/Entity/User.php
5. src/Entity/Operator.php
6. src/Entity/Metadata.php

We'll need to remove the direct usage of entities and replace them with the usage of DTOs and repositories.

Here's an example of how we can modify the Volume entity:

```php
src/Entity/Volume.php
<<<<<<< HEAD
use Doctrine\ORM\Mapping as ORM;
use Doctrine\Common\Collections\ArrayCollection;

#[ORM\Entity(repositoryClass: "App\Repository\VolumeRepository")]
class Volume
{
    #[ORM\Id, ORM\GeneratedValue, ORM\Column(type: "integer")]
    private $id;

    #[ORM\Column(type: "string", length: 255)]
    private $name;

    /**
     * @var ArrayCollection
     *
     * @ORM\OneToMany(targetEntity="File", mappedBy="volume")
     */
    private $files;

    public function __construct()
    {
        $this->files = new ArrayCollection();
    }

    // getters and setters
}
=======
use App\Domain\Entity\VolumeData;
use App\Repository\VolumeRepository;

class Volume
{
    private VolumeData $volumeData;
    private VolumeRepository $volumeRepository;

    public function __construct(VolumeData $volumeData, VolumeRepository $volumeRepository)
    {
        $this->volumeData = $volumeData;
        $this->volumeRepository = $volumeRepository;
    }

    // getters and setters
}
>>>>>>> updated
```

This is just an example for the Volume entity. Similar changes will be made to the Disk, File, User, Operator, and Metadata entities.

Please confirm if this is the approach you want to take.