aider: Implemented the `FormBuilderService` class that uses the `FormBuilderConfigurationReader` to read the configuration and create a form for each entry under the forms key.
# Aider chat conversation:

USER: let's implement a new feature, let's create a form builder service that will take the configuration from `FormBuilderConfigurationReader` and create a form for each entry under forms key

ASSISTANT: To implement this feature, we will need to create a new service called `FormBuilderService`. This service will use the `FormBuilderConfigurationReader` to read the configuration and then create a form for each entry under the forms key.

Here are the steps we need to follow:

1. Create a new file `src/Service/FormBuilderService.php`.
2. In this file, we will define a class `FormBuilderService`.
3. This class will have a constructor that accepts an instance of `FormBuilderConfigurationReader`.
4. We will also define a method `buildForms` that will use the `FormBuilderConfigurationReader` to read the configuration and then create a form for each entry under the forms key.

Here is the code for the new service:

```php
src/Service/FormBuilderService.php
<<<<<<< HEAD
=======
<?php

namespace App\Service;

use Symfony\Component\Form\FormFactoryInterface;

class FormBuilderService
{
    private FormBuilderConfigurationReader $configReader;
    private FormFactoryInterface $formFactory;

    public function __construct(FormBuilderConfigurationReader $configReader, FormFactoryInterface $formFactory)
    {
        $this->configReader = $configReader;
        $this->formFactory = $formFactory;
    }

    public function buildForms(): array
    {
        $config = $this->configReader->parse();
        $forms = [];

        foreach ($config['forms'] as $formName => $formConfig) {
            $form = $this->formFactory->createNamed($formName, $formConfig['type'], null, $formConfig['options']);
            $forms[$formName] = $form;
        }

        return $forms;
    }
}
>>>>>>> updated
```

Please note that this code assumes that the configuration file has a specific structure. If the structure of the configuration file is different, the code in the `buildForms` method will need to be adjusted accordingly.