aider: Refactored DashboardController to include `name` and `path` properties for each route in the `routes` array.
# Aider chat conversation:

USER: in dashboard.html.twig are using this <a href="{{ path(route.name) }}">{{ route.name }}</a> but route.name doesn't exist so you need to find a way to fix and provide correct info

ASSISTANT: The `route.name` in the `dashboard.html.twig` file is referring to the name of each route in the `routes` array passed from the `DashboardController.php` file. If `route.name` doesn't exist, it means that the `routes` array doesn't have a `name` property for each route.

To fix this, we need to modify the `DashboardController.php` file to include the `name` property for each route. We can do this by creating a new array of routes where each route is an associative array with `name` and `path` properties.

Here are the steps to do this:

1. Modify the `DashboardController.php` file to create a new array of routes with `name` and `path` properties.

Here is the *edit block* for this change:

```php
src/Controller/Admin/DashboardController.php
<<<<<<< HEAD
        $routes = $this->router->getRouteCollection()->all();

        // Filter and group the routes as needed
        foreach ($routes as $routeName => $route) {
            if (str_starts_with($routeName, '_profiler') || str_starts_with($routeName, '_wdt') || str_starts_with($routeName, '_preview_error')) {
                unset($routes[$routeName]);
            }
        }
=======
        $allRoutes = $this->router->getRouteCollection()->all();
        $routes = [];

        // Filter and group the routes as needed
        foreach ($allRoutes as $routeName => $route) {
            if (str_starts_with($routeName, '_profiler') || str_starts_with($routeName, '_wdt') || str_starts_with($routeName, '_preview_error')) {
                continue;
            }

            $routes[] = [
                'name' => $routeName,
                'path' => $this->router->generate($routeName),
            ];
        }
>>>>>>> updated
```

This change will create a new array of routes where each route is an associative array with `name` and `path` properties. The `name` property is the name of the route and the `path` property is the generated URL for the route. The `dashboard.html.twig` file can then use these properties to display the list of routes.