aider: Add `status` property of type `PostStatus` enum to `Page` entity.
# Aider chat conversation:

USER: let's create a new feature for PostEntity for the status let's create an enum that would return an integer for the following enum statuses inherit, auto-draft, trash, private, pending, draft, future, publish, unlisted, remember we are using symfony 6 and php 8.2

ASSISTANT: To implement this feature, we need to create a new file for the `PostStatus` enum and modify the `PostEntity` class to include the `status` property.

Here are the steps:

1. Create a new file `src/Entity/PostStatus.php` for the `PostStatus` enum.
2. Modify the `src/Entity/PostEntity.php` file to include the `status` property, along with its getter and setter methods.

Let's start with the creation of the `PostStatus` enum. Here is the code for the new
^C KeyboardInterrupt

USER: why do we need to create a PostEntity ?

ASSISTANT: I apologize for any confusion. The `PostEntity` class should already exist in your project as it represents the `Post` entity in your database. We don't need to create it, but we need to modify it to include the new `status` property.

The `
^C KeyboardInterrupt

USER: the change should apply for `src/Entity/Page.php`

ASSISTANT: I see, you want to add the `status` property to the `Page` entity, not the `PostEntity`. My apologies for misunderstanding.

We will need to modify the `src/Entity/Page.php` file to include the `status` property, which will be of type `PostStatus` enum.

Here are the steps:

1. Create a new file `src/Entity/PostStatus.php` for the `PostStatus` enum.
2. Modify the `src/Entity/Page.php` file to include the `status` property, along with its getter and setter methods.

Let's start with the creation of the `PostStatus` enum. Here is the code for the new file:

```php
src/Entity/PostStatus.php
<<<<<<< HEAD
=======
<?php

namespace App\Entity;

enum PostStatus: int
{
    case INHERIT = 0;
    case AUTO_DRAFT = 1;
    case TRASH = 2;
    case PRIVATE = 3;
    case PENDING = 4;
    case DRAFT = 5;
    case FUTURE = 6;
    case PUBLISH = 7;
    case UNLISTED = 8;
}
>>>>>>> updated
```

Next, we will modify the `src/Entity/Page.php` file to include the `status` property, along with its getter and setter methods:

```php
src/Entity/Page.php
<<<<<<< HEAD
    #[ORM\Column(type: Types::SMALLINT)] public ?int $status = null,
=======
    #[ORM\Column(type: 'PostStatus')] public ?PostStatus $status = null,
>>>>>>> updated
```

```php
src/Entity/Page.php
<<<<<<< HEAD
    ) {

    }
=======
    ) {}

    public function getStatus(): ?PostStatus
    {
        return $this->status;
    }

    public function setStatus(?PostStatus $status): self
    {
        $this->status = $status;

        return $this;
    }
>>>>>>> updated
```

Please note that the `PostStatus` enum is a feature of PHP 8.1 and later. Make sure your environment supports this version of PHP.