aider: Added save and delete methods to PageRepository.
# Aider chat conversation:

USER: please update `PageRepository` and add require methods save and delete

ASSISTANT: To add the `save` and `delete` methods to the `PageRepository`, we need to modify the `src/Repository/PageRepository.php` file.

Here are the changes:

```php
src/Repository/PageRepository.php
<<<<<<< HEAD
//    public function findOneBySomeField($value): ?Page
//    {
//        return $this->createQueryBuilder('p')
//            ->andWhere('p.exampleField = :val')
//            ->setParameter('val', $value)
//            ->getQuery()
//            ->getOneOrNullResult()
//        ;
//    }
=======
    public function save(Page $page): void
    {
        $this->_em->persist($page);
        $this->_em->flush();
    }

    public function delete(Page $page): void
    {
        $this->_em->remove($page);
        $this->_em->flush();
    }
>>>>>>> updated
```

Please note that `_em` is the EntityManager instance, which is a property of the `ServiceEntityRepository` class that `PageRepository` extends.