aider: Added `save` and `delete` methods to `PageMetaRepository`.
# Aider chat conversation:

USER: let's add crud methods in `PageMetaRepository`

ASSISTANT: Sure, we can add the `save` and `delete` methods to the `PageMetaRepository` class.

Here are the changes:

```php
src/Repository/PageMetaRepository.php
<<<<<<< HEAD
    public function __construct(ManagerRegistry $registry)
    {
        parent::__construct($registry, PageMeta::class);
    }
=======
    public function __construct(ManagerRegistry $registry)
    {
        parent::__construct($registry, PageMeta::class);
    }

    public function save(PageMeta $pageMeta): void
    {
        $this->_em->persist($pageMeta);
        $this->_em->flush();
    }

    public function delete(PageMeta $pageMeta): void
    {
        $this->_em->remove($pageMeta);
        $this->_em->flush();
    }
>>>>>>> updated
```

This will add the `save` and `delete` methods to the `PageMetaRepository` class. The `save` method will persist the `PageMeta` entity and flush the entity manager. The `delete` method will remove the `PageMeta` entity and flush the entity manager.