aider: Refactored `BaseController` to add a `handleFormSubmission` method that returns a callable if the form is submitted and valid.
# Aider chat conversation:

USER: let's refactor and implement a new feature in `src/Controller/BaseController.php` if the form is submited and valid to return a callable

ASSISTANT: Sure, we can add a new method to the `BaseController` class that will handle form submission and validation. This method will return a callable that can be used in your controllers.

Here are the changes:

```php
src/Controller/BaseController.php
<<<<<<< HEAD
class BaseController extends AbstractController
{
    // Add common methods for controllers here
}
=======
class BaseController extends AbstractController
{
    // Add common methods for controllers here

    protected function handleFormSubmission(FormInterface $form, callable $onSuccess): callable
    {
        return function (Request $request) use ($form, $onSuccess) {
            $form->handleRequest($request);

            if ($form->isSubmitted() && $form->isValid()) {
                return $onSuccess($form->getData());
            }

            return null;
        };
    }
}
>>>>>>> updated
```

This will add the `handleFormSubmission` method to the `BaseController` class. This method takes a `FormInterface` instance and a callable `$onSuccess` function as parameters. It returns a callable that handles the form submission and validation. If the form is submitted and valid, it calls the `$onSuccess` function with the form data. Otherwise, it returns `null`.